pyproject.toml
[project]
name = "lst-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-anthropic>=1.3.4",
    "langchain-core>=1.2.18",
    "langchain-openai>=1.1.11",
    "langgraph>=1.1.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "z3-solver>=4.16.0.0",
]
